server-src/planner.c: typo.
git-svn-id: https://amanda.svn.sourceforge.net/svnroot/amanda/amanda/trunk@3111 a8d146d6-cc15-0410-8900-af154a0219e0

+++ server-src/planner.c
@@ -2217,8 +2217,8 @@ static void analyze_estimate(
 	    if (est(dp)->estimate[i].level >= 0) {
 		g_fprintf(stderr,("    level: %d  nsize: %lld csize: %lld\n"),
 			  est(dp)->estimate[i].level,
-			  est(dp)->estimate[i].nsize,
-			  est(dp)->estimate[i].csize);
+			  (long long)est(dp)->estimate[i].nsize,
+			  (long long)est(dp)->estimate[i].csize);
 	    }
 	}
 	log_add(L_WARNING, _("%s %s %s 0 %s"), dp->host->hostname, qname,
